In this commit, function parameters are long double****

+++ 2nd/common-globals.h
@@ -1,8 +1,8 @@
-extern double alpha;
-extern double omega;
-extern double mass;
+extern long double alpha;
+extern long double omega;
+extern long double mass;
 extern int number_of_mesh;
-extern double epsilon;
-extern double offset;
-extern double low_lim;
-extern double up_lim;
+extern long double epsilon;
+extern long double offset;
+extern long double low_lim;
+extern long double up_lim;
